Move generic utils out of storage code to utils.math
Moved to `utils.math`

- valid_mask
- is_almost_int
