+++ tests/test_utils.py
@@ -20,7 +20,7 @@ from hypothesis.strategies import integers, text
 from pandas import to_datetime
 
 from datacube.helpers import write_geotiff
-from datacube.utils.math import num2numpy
+from datacube.utils.math import num2numpy, is_almost_int, valid_mask
 from datacube.model import MetadataType
 from datacube.model.utils import xr_apply, traverse_datasets, flatten_datasets, dedup_lineage
 from datacube.testutils import mk_sample_product, make_graph_abcde, gen_dataset_test_dag, dataset_maker
@@ -799,6 +799,42 @@ def sample_document_files(data_folder):
              for f, num_docs in files]
 
     return files
+
+
+def test_is_almost_int():
+    assert is_almost_int(1, 1e-10)
+    assert is_almost_int(1.001, .1)
+    assert is_almost_int(2 - 0.001, .1)
+    assert is_almost_int(-1.001, .1)
+
+
+def test_valid_mask():
+    xx = np.zeros((4, 8), dtype='float32')
+    mm = valid_mask(xx, 0)
+    assert mm.dtype == 'bool'
+    assert mm.shape == xx.shape
+    assert not mm.all()
+    assert not mm.any()
+
+    mm = valid_mask(xx, 13)
+    assert mm.dtype == 'bool'
+    assert mm.shape == xx.shape
+    assert mm.all()
+
+    mm = valid_mask(xx, None)
+    assert mm.dtype == 'bool'
+    assert mm.shape == xx.shape
+    assert mm.all()
+
+    mm = valid_mask(xx, np.nan)
+    assert mm.dtype == 'bool'
+    assert mm.shape == xx.shape
+    assert mm.all()
+
+    xx[0, 0] = np.nan
+    mm = valid_mask(xx, np.nan)
+    assert not mm[0, 0]
+    assert mm.sum() == (4*8-1)
 
 
 def test_num2numpy():
